requirements.txt
discord.py>=2.3
fastapi>=0.110
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.5
SQLAlchemy>=2.0
python-dotenv>=1.0
aiosqlite>=0.20
